got click to convert to correct array location, organized logic and files, outlined next steps pseudocode
